refactor: all screens in subfolders
