+++ App.tsx
@@ -5,26 +5,21 @@ import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import { createStackNavigator } from "@react-navigation/stack";
 import Ionicons from "@expo/vector-icons/Ionicons";
 
-import HomeScreen from "./src/screens/HomeScreen";
-import MomsOverviewScreen from "./src/screens/MomsOverviewScreen";
-import CalendarScreen from "./src/screens/CalendarScreen";
-import SessionDetailsScreen from "./src/screens/SessionDetailsScreen";
-import CourseDetailsScreen from "./src/screens/CourseDetailsScreen";
-import MomDetailsScreen from "./src/screens/MomDetailsScreen";
-import MomAddScreen from "./src/screens/MomAddScreen";
-import CourseOverviewScreen from "./src/screens/CoursesOverviewScreen";
-import CourseAddScreen from "./src/screens/CourseAddScreen";
-import SessionAddScreen from "./src/screens/SessionAddScreen";
+import MomsOverviewScreen from "./src/screens/moms/MomsOverviewScreen";
+import CalendarScreen from "./src/screens/calendar/CalendarScreen";
+import SessionDetailsScreen from "./src/screens/calendar/SessionDetailsScreen";
+import CourseDetailsScreen from "./src/screens/courses/CourseDetailsScreen";
+import MomDetailsScreen from "./src/screens/moms/MomDetailsScreen";
+import MomAddScreen from "./src/screens/moms/MomAddScreen";
+import CourseOverviewScreen from "./src/screens/courses/CoursesOverviewScreen";
+import CourseAddScreen from "./src/screens/courses/CourseAddScreen";
+import SessionAddScreen from "./src/screens/calendar/SessionAddScreen";
 import { RootStackParamList } from "./src/types/RootStackParamListType";
 import { Amplify } from "aws-amplify";
 import {
   withAuthenticator,
   useAuthenticator,
 } from "@aws-amplify/ui-react-native";
-import amplifyconfig from "./src/amplifyconfiguration.json";
-// Amplify.configure(amplifyconfig);
-
-// Imports the awsExports configuration file generated by the Amplify CLI. This file contains the AWS service configurations (like Cognito, AppSync, etc.) specific to your project.
 import awsExports from "./src/aws-exports";
 
 Amplify.configure(awsExports);
@@ -114,8 +109,6 @@ const App = () => {
   const Tab = createBottomTabNavigator();
   return (
     <>
-      {/* <Authenticator.Provider>
-          <Authenticator> */}
       <SafeAreaView style={styles.topSafeArea} />
       <SafeAreaView style={styles.bottomSafeArea}>
         <NavigationContainer>
@@ -141,11 +134,6 @@ const App = () => {
               tabBarInactiveTintColor: "gray",
             })}
           >
-            {/* <Tab.Screen
-                name="Home"
-                component={HomeScreen}
-                options={{ headerShown: false }}
-              /> */}
             <Tab.Screen
               name="Kalender"
               component={CalendarStack}
@@ -164,8 +152,6 @@ const App = () => {
           </Tab.Navigator>
         </NavigationContainer>
       </SafeAreaView>
-      {/* </Authenticator>
-        </Authenticator.Provider> */}
     </>
   );
 };
