--- src/screens/HomeScreen.tsx
@@ -1,20 +1,0 @@
-import React from "react";
-import { StyleSheet, Text, View } from "react-native";
-
-const HomeScreen = () => {
-  return (
-    <View style={styles.container}>
-      <Text>Home!</Text>
-    </View>
-  );
-};
-
-export default HomeScreen;
-
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: "#f0f0f0",
-    flex: 1,
-    paddingTop: 15,
-  },
-});
